LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/button-layout/display-other-expected.txt

after

FAIL display: block assert_equals: buttonTop expected 0 but got 2
FAIL display: run-in assert_true: display: run-in is not supported expected true got false
FAIL display: flow assert_true: display: flow is not supported expected true got false
PASS display: flow-root
PASS display: table
PASS display: list-item
FAIL display: table-row-group assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-header-group assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-footer-group assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-row assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-cell assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-column-group assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-column assert_equals: buttonLeft expected 120 but got 122
FAIL display: table-caption assert_equals: buttonLeft expected 120 but got 122
FAIL display: ruby-base assert_true: display: ruby-base is not supported expected true got false
FAIL display: ruby-text assert_true: display: ruby-text is not supported expected true got false
FAIL display: ruby-base-container assert_true: display: ruby-base-container is not supported expected true got false
FAIL display: ruby-text-container assert_true: display: ruby-text-container is not supported expected true got false
